+++ R/fscores.internal.R
@@ -4,20 +4,24 @@ setMethod(
 	definition = function(object, rotate = '', full.scores = FALSE, method = "EAP",
                           quadpts = NULL, response.pattern = NULL, theta_lim, MI, 
 	                      returnER = FALSE, verbose = TRUE, gmean, gcov, scores.only,
-	                      full.scores.SE)
+	                      full.scores.SE, return.acov = FALSE)
 	{
 	    #local functions for apply
-	    MAP <- function(ID, scores, pars, tabdata, itemloc, gp, prodlist, CUSTOM.IND){
+	    MAP <- function(ID, scores, pars, tabdata, itemloc, gp, prodlist, CUSTOM.IND,
+	                    return.acov = FALSE){
 	        estimate <- try(nlm(MAP.mirt,scores[ID, ],pars=pars, patdata=tabdata[ID, ],
 	                            itemloc=itemloc, gp=gp, prodlist=prodlist, hessian=TRUE, 
                                 CUSTOM.IND=CUSTOM.IND))
 	        if(is(estimate, 'try-error'))
 	            return(rep(NA, ncol(scores)*2))
-	        SEest <- try(sqrt(diag(solve(estimate$hessian))))
+            vcov <- try(solve(estimate$hessian))
+            if(return.acov) return(vcov)
+	        SEest <- try(sqrt(diag(vcov)))
 	        if(is(SEest, 'try-error')) SEest <- rep(NA, ncol(scores))
 	        return(c(estimate$estimate, SEest))
 	    }
-	    ML <- function(ID, scores, pars, tabdata, itemloc, gp, prodlist, CUSTOM.IND){
+	    ML <- function(ID, scores, pars, tabdata, itemloc, gp, prodlist, CUSTOM.IND,
+	                   return.acov = FALSE){
 	        if(any(scores[ID, ] %in% c(-Inf, Inf)))
                 return(c(scores[ID, ], rep(NA, ncol(scores))))
 	        estimate <- try(nlm(MAP.mirt,scores[ID, ],pars=pars,patdata=tabdata[ID, ],
@@ -25,11 +29,14 @@ setMethod(
                                 CUSTOM.IND=CUSTOM.IND))
 	        if(is(estimate, 'try-error'))
 	            return(rep(NA, ncol(scores)*2))
-	        SEest <- try(sqrt(diag(solve(estimate$hessian))))
+	        vcov <- try(solve(estimate$hessian))
+	        if(return.acov) return(vcov)
+	        SEest <- try(sqrt(diag(vcov)))
 	        if(is(SEest, 'try-error')) SEest <- rep(NA, ncol(scores))
 	        return(c(estimate$estimate, SEest))
 	    }
-	    WLE <- function(ID, scores, pars, tabdata, itemloc, gp, prodlist, CUSTOM.IND){
+	    WLE <- function(ID, scores, pars, tabdata, itemloc, gp, prodlist, CUSTOM.IND,
+                        return.acov = FALSE){
 	        estimate <- try(nlm(gradnorm.WLE,scores[ID, ],pars=pars,patdata=tabdata[ID, ],
 	                            itemloc=itemloc, gp=gp, prodlist=prodlist, CUSTOM.IND=CUSTOM.IND))
 	        if(is(estimate, 'try-error'))
@@ -37,16 +44,37 @@ setMethod(
             TI <- 0
             for(i in 1L:(length(itemloc)-1L))
                 TI <- TI + iteminfo(pars[[i]], Theta=estimate$estimate)
+            if(return.acov) return(1/TI)
 	        SEest <- 1 / sqrt(TI)
 	        return(c(estimate$estimate, SEest))
 	    }
-	    EAP <- function(ID, log_itemtrace, tabdata, ThetaShort, W){
+	    EAP <- function(ID, log_itemtrace, tabdata, ThetaShort, W, return.acov = FALSE){
+            nfact <- ncol(ThetaShort)
 	        L <- rowSums(log_itemtrace[ ,as.logical(tabdata[ID,]), drop = FALSE])
 	        thetas <- colSums(ThetaShort * exp(L) * W / sum(exp(L) * W))
-	        SE <- sqrt(colSums(t((t(ThetaShort) - thetas))^2 * exp(L) * W / sum(exp(L) * W)))
+	        thetadif <- t((t(ThetaShort) - thetas))
+            Thetaprod <- matrix(0, nrow(ThetaShort), nfact * (nfact + 1L)/2L)
+            ind <- 1L
+            for(i in 1L:nfact){
+                for(j in 1L:nfact){
+                    if(i <= j){
+                        Thetaprod[,ind] <- thetadif[,i] * thetadif[,j]
+                        ind <- ind + 1L
+                    }
+                }
+            }
+            vcov <- matrix(0, nfact, nfact)
+            vcov[lower.tri(vcov, TRUE)] <- colSums(Thetaprod * exp(L) * W / sum(exp(L) * W))
+            if(nfact > 1L) vcov <- vcov + t(vcov) - diag(diag(vcov))
+            if(return.acov) return(vcov)
+	        SE <- sqrt(diag(vcov))
 	        return(c(thetas, SE))
 	    }
-
+        
+        if(return.acov && MI != 0)
+            stop('simultaneous impute and return.acov option not supported')
+	    if(return.acov && returnER)
+	        stop('simultaneous returnER and return.acov option not supported')
         if(!is.null(response.pattern)){
             drop <- FALSE
             if(!is.matrix(response.pattern)){
@@ -64,7 +92,7 @@ setMethod(
                            technical=list(customK=object@K))
             ret <- fscores(newmod, rotate=rotate, full.scores=full.scores, scores.only=scores.only,
                            method=method, quadpts=quadpts, verbose=FALSE,
-                           response.pattern=NULL)
+                           response.pattern=NULL, return.acov=return.acov)
             if(!scores.only || !full.scores)
                 ret[,1L:ncol(response.pattern)] <- ret[,1L:ncol(response.pattern)] +
                     matrix(mins, nrow(ret), ncol(response.pattern), byrow=TRUE)
@@ -133,19 +161,22 @@ setMethod(
                 itemtrace <- computeItemtrace(pars=pars, Theta=Theta, itemloc=itemloc, 
                                               CUSTOM.IND=CUSTOM.IND)
                 log_itemtrace <- log(itemtrace)
-        	    tmp <- myApply(X=matrix(1L:nrow(scores)), MARGIN=1L, FUN=EAP, log_itemtrace=log_itemtrace,
-                               tabdata=tabdata, ThetaShort=ThetaShort, W=W)
-        	    scores <- tmp[ ,1:nfact, drop = FALSE]
-        	    SEscores <- tmp[ ,-c(1:nfact), drop = FALSE]
+                if(method == 'EAP' && return.acov){
+                    tmp <- myApply(X=matrix(1L:nrow(scores)), MARGIN=1L, FUN=EAP, log_itemtrace=log_itemtrace,
+                                   tabdata=tabdata, ThetaShort=ThetaShort, W=W, return.acov=TRUE)
+                } else {
+            	    tmp <- myApply(X=matrix(1L:nrow(scores)), MARGIN=1L, FUN=EAP, log_itemtrace=log_itemtrace,
+                                   tabdata=tabdata, ThetaShort=ThetaShort, W=W)
+            	    scores <- tmp[ ,1:nfact, drop = FALSE]
+            	    SEscores <- tmp[ ,-c(1:nfact), drop = FALSE]                
+                }
             }
     		if(method == "EAP"){
                 #do nothing
     		} else if(method == "MAP"){
                 tmp <- myApply(X=matrix(1L:nrow(scores)), MARGIN=1L, FUN=MAP, scores=scores, pars=pars,
                                tabdata=tabdata, itemloc=itemloc, gp=gp, prodlist=prodlist, 
-                               CUSTOM.IND=CUSTOM.IND)
-                scores <- tmp[ ,1:nfact, drop = FALSE]
-                SEscores <- tmp[ ,-c(1:nfact), drop = FALSE]
+                               CUSTOM.IND=CUSTOM.IND, return.acov=return.acov)
     		} else if(method == "ML"){
                 isna <- apply(object@tabdata[,-ncol(object@tabdata)], 1L, 
                               function(x) sum(is.na(x)))[keep]
@@ -157,9 +188,7 @@ setMethod(
                 SEscores[allzero,] <- NA
                 tmp <- myApply(X=matrix(1L:nrow(scores)), MARGIN=1L, FUN=ML, scores=scores, pars=pars,
                                tabdata=tabdata, itemloc=itemloc, gp=gp, prodlist=prodlist,
-                               CUSTOM.IND=CUSTOM.IND)
-                scores <- tmp[ ,1:nfact, drop = FALSE]
-                SEscores <- tmp[ ,-c(1:nfact), drop = FALSE]
+                               CUSTOM.IND=CUSTOM.IND, return.acov=return.acov)
     		} else if(method == 'WLE'){
                 if(nfact > 1L)
                     stop('WLE method only supported for unidimensional models')
@@ -168,16 +197,21 @@ setMethod(
                 tmp <- myApply(X=matrix(1L:nrow(scores)), MARGIN=1L, FUN=WLE, scores=scores, pars=pars,
                                tabdata=tabdata, itemloc=itemloc, gp=gp, prodlist=prodlist, 
                                CUSTOM.IND=CUSTOM.IND)
-                scores <- tmp[ ,1:nfact, drop = FALSE]
-                SEscores <- tmp[ ,-c(1:nfact), drop = FALSE]
             } else {
                 stop('method not defined')
             }
-    		colnames(scores) <- paste('F', 1:ncol(scores), sep='')
-            if(impute){
-                list_SEscores[[mi]] <- SEscores
-                list_scores[[mi]] <- scores
-            } 
+    		if(return.acov){
+    		    scores <- tmp
+                if(nrow(scores) < ncol(scores)) scores <- t(scores)
+    		} else {
+    		    scores <- tmp[ ,1:nfact, drop = FALSE]
+    		    SEscores <- tmp[ ,-c(1:nfact), drop = FALSE]
+    		    colnames(scores) <- paste('F', 1:ncol(scores), sep='')
+    		    if(impute){
+    		        list_SEscores[[mi]] <- SEscores
+    		        list_scores[[mi]] <- scores
+    		    }
+    		}
         }
         if(impute){
             scores <- list_scores[[1L]]/MI
@@ -198,6 +232,13 @@ setMethod(
                 stabdata2 <- apply(tabdata2, 1, paste, sep='', collapse = '/')
                 sfulldata <- apply(object@fulldata, 1, paste, sep='', collapse = '/')
                 scoremat <- scores[match(sfulldata, stabdata2), , drop = FALSE]
+                if(return.acov){
+                    ret <- vector('list', nrow(scoremat))
+                    for(i in 1L:nrow(scoremat))
+                        ret[[i]] <- matrix(scoremat[i,], nfact, nfact)
+                    names(ret) <- 1:nrow(scoremat)
+                    return(ret)
+                }
                 SEscoremat <- SEscores[match(sfulldata, stabdata2), , drop = FALSE]
     			colnames(scoremat) <- colnames(scores)
     			colnames(SEscoremat) <- paste0('SE_',colnames(scores))
@@ -211,6 +252,13 @@ setMethod(
             if(scores.only) return(scoremat)
 			else return(cbind(fulldata,scoremat))
 		} else {
+            if(return.acov){
+                ret <- vector('list', nrow(scores))
+                for(i in 1L:nrow(scores))
+                    ret[[i]] <- matrix(scores[i,], nfact, nfact)
+                names(ret) <- paste0('pattern_', 1:nrow(scores))
+                return(ret)
+            }
             r <- object@tabdata[,ncol(object@tabdata)]
             T <- E <- matrix(NA, 1, ncol(scores))
             for(i in 1L:nrow(scores)){
@@ -243,13 +291,13 @@ setMethod(
 	definition = function(object, rotate = '', full.scores = FALSE, method = "EAP",
 	                      quadpts = NULL, response.pattern = NULL, theta_lim, MI,
 	                      returnER = FALSE, verbose = TRUE, gmean, gcov, scores.only,
-	                      full.scores.SE)
+	                      full.scores.SE, return.acov = FALSE)
 	{
         class(object) <- 'ExploratoryClass'
         ret <- fscores(object, rotate = 'CONFIRMATORY', full.scores=full.scores, method=method, quadpts=quadpts,
                        response.pattern=response.pattern, returnER=returnER, verbose=verbose,
                        mean=gmean, cov=gcov, scores.only=scores.only, theta_lim=theta_lim, MI=MI,
-                       full.scores.SE=full.scores.SE)
+                       full.scores.SE=full.scores.SE, return.acov = return.acov)
         return(ret)
 	}
 )
@@ -261,7 +309,7 @@ setMethod(
     definition = function(object, rotate = '', full.scores = FALSE, method = "EAP",
                           quadpts = NULL, response.pattern = NULL, theta_lim, MI,
                           returnER = FALSE, verbose = TRUE, gmean, gcov, scores.only,
-                          full.scores.SE)
+                          full.scores.SE, return.acov = FALSE)
     {
         cmods <- object@cmods
         ngroups <- length(cmods)
@@ -276,9 +324,22 @@ setMethod(
             ret[[g]] <- fscores(cmods[[g]], rotate = 'CONFIRMATORY', full.scores=full.scores, method=method,
                            quadpts=quadpts, returnER=returnER, verbose=verbose, theta_lim=theta_lim,
                                 mean=gmean[[g]], cov=gcov[[g]], scores.only=FALSE, MI=MI,
-                           full.scores.SE=full.scores.SE)
+                           full.scores.SE=full.scores.SE, return.acov=return.acov)
         names(ret) <- object@groupNames
         if(full.scores){
+            if(return.acov){
+                group <- object@group
+                groupNames <- object@groupNames
+                count <- numeric(length(groupNames))
+                out <- vector('list', length(group))
+                for(i in 1L:length(group)){
+                    which <- which(groupNames %in% group[i])
+                    count[which] <- count[which] + 1L
+                    out[[i]] <- ret[[which]][[count[which]]]
+                }
+                names(out) <- 1L:length(out)
+                return(out)
+            }
             id <- c()
             fulldata <- matrix(NA, 1, ncol(ret[[1]]))
             for(g in 1L:ngroups){